Fix ``sff_extract`` tool, drop type annotation
I don't think this tool is really used much, but it is loaded in
`tool_conf.xml`.
https://github.com/galaxyproject/galaxy/pull/11877 added type
annotations, and that's a syntax error on 2.7

The alternative would be to fix the rest of the tool for 3.x,
but I think we're just wasting our time there and we shouldn't
touch this again.