Drop batch middleware
I can't find a trace of this being used in the client. We should
re-implement it as a ASGI middleware if we've identified a need for it,
but I don't love how this makes the call a potentially very long
blocking call. To fetch many items in batch we might be better off using
Server Sent Events, which can trickle out intermediate updates.