Better binary detection with FilePrefix & libmagic
FilePrefix reads the first Megabytes of a file, which is much more
than the previously used `check_binary` which read only 1KB.

Since we're re-using FilePrefix instance this shouldn't have any impact
on performace. In addition we're now using `python-magic` to detect
the encoding and mime-type. We're using the mime-type detection
for PDF only at this point, but we're also using the detected encoding
if we encounter non-utf8 content.