Rework workflow index API for more consistency between deleted/hidden and showing shared workflows.
So currently is show_deleted is true - you'll also non-deleted workflows that are shared with you but not your own non-deleted workflows. Likewise, regardless of show_hidden - yo
u always get the same bunch of shared workflows. Also none of that is documented in tests or the docstring.

This tries to fix that up a bit by making showing shared workflows a parameterized boolean that defaults in a consistent, safe way with the parameters show_deleted and show_hidden. It will default to true if those are false and to false if either of those are true - so it always builds a consistent set of workflows. It also gives very explicit and tested
 errors if invalid combinations are requested (e.g. explicitly asking for deleted workflows shared with you for instance).