Fix workflow bookmarking API.
And drop client-side hack that I assume was working around it. I tried dropping the tags argument that seems like shouldn't be needed and the results were not being persisted. Presumably someone realized the tags caused a flush and so sticking them in there worked around a bug related to a missing flush in this API.