ensure that Singularity doesn't run with HOME=/root
When Galaxy is run in Docker, it starts uwsgi and passes the `--uid` and `--gid` flags to make the daemon process drop privileges. This leaves the HOME environment variable unchanged, so typically HOME=/root. In this case, when `singularity build` runs, it looks in for the `$HOME/.docker/config.json` file which fails if HOME is inaccessible. The result is that the container does not get built and cached and rather gets rebuilt every time a tool runs.

This PR addresses that case by looking for an inaccessible HOME directory and choosing an alternative - either the value in GALAXY_HOME or /tmp.