Don't create a BGZF index when running the compressing converter
The index file is not used because it is generated again in `set_meta()`
for the `VcfGz` datatype, and ignored for all others.

Additionally, the index file is currently generated alongside the input
dataset, which fails when the input directory is not writeable (such as
when running in a container).

Fix https://github.com/galaxyproject/galaxy/issues/13916 .

Also, replace outdated `tabix` requirement with the latest `htslib`.