Fix contains_collection check for empty root collections
Comparing hdca_id with parent_id is just wrong, as hdca_id is the id for
a `HistoryDatasetCollectionAssociation`, while `parent_id` is the id for
a `DatasetCollection`. This happened to work regardless of the first if
check if the root collection contains at least one DCE, but that won't
work for collections without elements.