Raise error when bulk undeleting
Instead of doing nothing and treating the item as successful, we now raise a specific error.
This helps to determine if the bulk operation will trigger a history update when an item succeeds.
Otherwise, the UI will keep waiting for a history update that may never occur.
For the same reason, when purging an already purged item, we will touch the `update_time` but do nothing, so we notify the history to expect a *change*.