Relocate dynamic outputs for CWL jobs
Add command to run a relocate script to the `build_command()` function.

Also update the `handle_outputs()` function (used by the relocate script)
to now use a CWL-specific `params.json` file, since the one created by
`setup_external_metadata()` is not yet available to the relocate script
if the metadata is set within a celery task.