Fix bug in display_data in spaln.py and blast.py
The correct arg name is `dataset`, not `data`. The method body contains
a call to the same method in the parent class (defined in data.py),
passing `data` (i.e., dataset) as `data=data`, and that doesn't work,
because the parent method expexts `dataset`. Renaming the argument fixes
this error + the typing error that would've happened when
adding type hints.