Rename parameter in display_data in data.py (see note)
Change from "data" to "dataset".
Use of this parameter is inconsistent across subclasses of Data when
they invoke the parent class method: some pass the HDA as "data", some
as "dataset"; as a result, values passed as "dataset" were lost.

I chose "dataset" becuase it's more consistent with how we typically refer to an
HDA argument in the code base. It may have been more accurate to use
"hda" instead, but there are numerous similar cases that would need to
be renamed following this pattern for consistency (even across the
datatypes directory).