Dereference symlinks in from_work_dir handling
That's needed so we can collect outputs from a symlink. It comes in
handy (see the previous commit) when a tool author wants to name an
output e.g after a element_identifier. A common pattern is then to use
a `cp` to copy the output, but that may not work if Galaxy is stopping
the process and the `&&` portion isn't guaranteed to run.