Use dataset uuid to authorize history export archive downloads
It's a pretty common thing that users try to import history archives to
another server via the URL we provide, or try to download the archive
via cURL or friends. You just need to know and remember to publish the
history for that to work. This instead uses the URL itself to authorize
access.

The history id and jeha id are kind-of predictable
(already part of the url) since they're just hashed integers, but we can
include the uuid, which isn't easily proedictable. That means anyone
with access to the URL can download that particular history archive,
which I think is the expectation.

Fixes https://github.com/galaxyproject/galaxy/issues/14447 and part of https://github.com/galaxyproject/galaxy/issues/12761