Allow router to reactively control the masthead.
The goal here is to replace the hacks in the workflow centric UI where we want a completely different masthead depending on the router we're rendering. This just sets up the plumbing though and uses restoring the highlight of the current masthead tab selected from older Galaxies with the new version of the masthead.