Fix post job action getting lost when node is made active
And many more reactivity fixes.

The core of the issue was that the FormElement would emit a onUpdate
event, which triggers a request to `api/workflows/build_module`.
The FormSection that controls the post job action data is not ready at
this point, so when the data comes back in we have to break the
connection because the post job action that sets the datatype is
missing.

We really don't need to call `build_module` though when change post job
actions, they don't require the backend at all, we just need to bubble
them up to the Index component, which can set the postJobAction on the
active node (let's ignore that we should manipulate the data, not the vue instance
...).

In order for that to be properly reactive I had to remove the majority
of the getNode hacks.

We do still rely on this in the Terminals and Connectors, so those can
get out of sync still, but it fixes an annoying issue where connections
are dropped without notice.