make wait_on use real time (not accumulated sleep time)
the function that "realises waiting for a result or timeout"
currently accumulated the sleep times and compared it to the
timeout. this leads to unexpected long test timeouts
(approx 1.5 - 2 times .. independent of the results of the
previous commit: 9006994b46)

Unexpected in the sense of the planemo CLI docs:

```
Maximum runtime of a single test in seconds.
```

Question is if in one of the multiple places where the function
is called accumulated sleep times are really expected (and maybe
we can just change the timeout at these places).

Also removed the sleep_ parameter which is apparently unused
(and would have complicated importing).