Handle errors gracefully with manage_db.sh
The `--raiseerr` option (False by default) will print a full stack trace
in case of an error. Omitting it will print out a friendly log message
instead. However, this option does not exist in the context of the
manage_db.sh script, which resulted in an ugly traceback. This fixes it:
if a command fails, the script will print the friendly log message.