Replace ``unittest.TestCase`` in API, framework, performance, Selenium and ToolShed tests
In API tests, replace no more working `uses_test_history` decorator (because
the `history_id` method parameter is interpreted as a fixture) with a real
class-level pytest fixture.
As a consequence, replace use of `self.history_id` in API tests because it
interferes with the class-level `history_id` fixture.

Also:
- Some type annotations