Make dataset parameter required
It's never None for this class, so having the parameter as optional is
semantically incorrect. This, however, does violate the Liskov
substitution principle (subclass method signature more strict than
parent definition), but we can fix this (as well as many other such
cases in datatypes) by slightly modifying the object hierarchy. Should
be submitted in a follow-up PR.