Fix loading subworkflows from editor interface
This probably broke in
https://github.com/galaxyproject/galaxy/pull/13984. There were 3 issues
here: Subworkflows are referenced by Workflow.id, not StoredWorkflow.id,
so we'd build the wrong url. The other issue is that if we push the new url
into the router nothing happens as we re-use the component. Instead we
need to listen to route changes and reload the editor config.
The third problem is that the main editor component loads the workflow
data only in the created hook.

I think this needs further refactoring to not use the controller
endpoint; everything is available through exisiting APIs and can be
cached (toolbox, workflows, tags), but this fixes subworkflow editing
for now.