Fix broken subworkflow reodering on workflow import
We started reordering workflow steps on import in
https://github.com/galaxyproject/galaxy/pull/13641.

If subworkflow steps are reordered before the input connections are
set the wrong subworkflow steps are referenced when we build the
workflow connection. This broke the IWC tests for gromacs.

We now do not reorder subworkflow steps until the parent workflow is
being reordered, which is after all connections are set up.