Create traceback.txt when external metadata fails
It's always a little tricky to debug this, and often you'll have to run
the metadata script by hand activating galaxy's virtualenv and setting
PYTHONPATH to galaxy's lib folder. This should improve the situation a
lot, both in production and when debugging tests.

Should help (a little) with
https://github.com/galaxyproject/galaxy/issues/14931 (but this is in
celery, so we wouldn't pass through this code).