accept tool execution exception if expect_failure
some CI tool tests at IUC which had `expect_failure="true"`
were failing "suddenly". I guess this coincides with the
introducion of planemo 0.75.x. But I'm not sure if this is
due to a change in planemo (gravity), the testing function
in `tool-util`, or changes to the paramater validators.

I guess its one of the latter two, but I'm not sure why
we did not see this earlier. Also I'm not sure if this
is the correct fix.

xref https://github.com/galaxyproject/tools-iuc/pull/4910