Move test_dbscript into app
JUSTIFICATION:
As per @jmchilton:
The package tests should pretend they don't have galaxy_directory()
available because they're meant to be installed into a Python env.
This test is fundamentally not testing anything about the Galaxy data
package - it is testing scripts placed in Galaxy root directory and that
would not be available in an installed Galaxy environment.