shorten long text in content assertion output
I had a case where the assertion was applied on a very long file.
This just flooded my terminal which was not helpful during debugging.
So I thought maybe a shortened output (like for compare="diff")
in case of long data maybe usefull.

Output format and length threshold are inspiered by the output for
compare="diff". Not entirely sure if optimal.