Add Protocol, type hints to generate_primary_file()
Also, change dataset parameter from optional to mandatory.

Justification:
1. The definition in a subclass in anvio.py violated Liskov substitution
   principle (parameter was not optinal)
2. Method is only called with a dataset argument present, so making the
   paramter optinal is unnecessary.

Why a Protocol instead of DatasetInstance, Bunch, etc?
The method is called with a Bunch, a Dataset(Bunch), or a
DatasetInstance. Rather than using a Union of types or creating a new
type, a Protocol seems the cleanest and most explicit way of stating
intent (i.e., we don't care what the type of the object is, but the
object must have these properties).