Do not OpenAPI schema lint for operation-4xx-response.
We have good infrastructure for error handling in the API and it could be improved in many ways - but I don't think keeping a list of exclusions is helpful. This is just going to get in the way and encourage unhelpful boilerplate IMO. This is the equivalent of Java checked exceptions - a bad idea.

xref: https://redocly.com/docs/cli/rules/operation-4xx-response/