Fix for local workflow file upload failure
A less trivial and hacky version of #15318 in response to improvement suggested by @mvdbeek 
## What:
fixed a bug - latest dev workflow upload form works for URLs but fails for local files. 

## Why: 

Try uploading a local workflow file from the workflow import form.
It breaks, complaining there's no appropriate method in the request. 
The logic in _WorkflowsAPIController.create_ wants "archive_source" to be set but empty for a local file.

## How to test the changes?

Adding the line suggested and rebuilding the client seems to fix it.

## License
- [x] I agree to license these and all my past contributions to the core galaxy codebase under the [MIT license](https://opensource.org/licenses/MIT).