Don't call setValue in FormElement setup
We already have the necessary data that were passed in via the prop.
Instead we emit the current form state once in the FormDisplay
component's `created` hook.  That works reliably, also if there is no
acutal parameter value.

To handle the no parameter value case in the workflow editor (as happens in the
`__BUILD_LIST__` tool) I've changed the initial empty Object value to null,
so that when we determine whether or not we have to update the module
we're not thrown off by the empty Object value.