Rewrite datasets cleanup API action
Purge associations and datasets in 2 steps:
First synchronously mark all associations as purged in one transaction.
Then purge internal datasets in a celery task or synchronously if not available.

This will ensure that the associations are marked as purged when the request finishes keeping the results consistent between subsequent requests.