Drop SentryWSGIMiddleware
It appears that we can't combine the WSGI middleware with the ASGI
middleware in the same app. We're currently not sending events that
occur in the WSGI portion of the app. With this change they're logged
again, though we do lose the WSGI request context (which contains,
browser, route, etc.) It's usually pretty clear from the traceback what
went wrong though, so I think this is much better than what's happening
now.