Use new session in db heartbeat
Both methods may be called from multiple threads. Each thread, when
using Galaxy's scoped session registry, reuses one Session instance. On
commit or rollback a session expires the ORM entities it contains. Thus,
if this method is called in the context of a thread that is using a
non-empty Session object, and we reuse the same session, it will be
committed or rolled back at the end of the method, thus espiring all its
contents, and causing errors elsewhere in the application. Using a new
Session instance solves this.