tests: enforce container type
this is probably not excactly what is done when listing containers
in the admin UI. This is because a container type is only set via
the destination a job is executed on. That is, for listing and
building both container types are allowed, which might lead to
unwanted outcomes if the container resolver config contains resolvers
for both container types (like the default does).