Revert pulsar-side stdio redirection
It just happened to pass tests in
https://github.com/galaxyproject/galaxy/pull/15907 because the test
framework looks at the stdio and stderr attributes, which are merged
from tool_std{out,err} and job_std{out,err}.
It however fails the tests for detect_errors that specifically checks
tool_stdout and tool_stderr.

I think it is a good idea to let pulsar handle the stdio redirection,
but for this to work we'd have to ship parts of the command line to pulsar
for final assembly, and that seems like a bigger change.