Add more fields to UpdateHistoryContentsPayload
Those fields are the most commonly used ones so it's worth defining them, the rest are dynamic for now.
Also, it requires to exclude the unset fields from the payload, otherwise those will be set to None by pydantic and the server logic will assume the user wants to set them to None.