Fix case insensitive orm filtering in lists
Not sure if this is the best location or way to fix it, but after `_split_filters` we won't have the `case_insensitive` flag available because the ParsedFilter wrapper gets dropped there.

Also we can be sure that if `case_insensitive` is True in that particular filter then we surely want to compare against the lower version of the column as the value to compare to already was converted to lower when parsing the filter.