Fix transaction error in unit test
In galaxy, HistoryAudit.prune() executes in the context of a separate thread, where it
starts and commits a new transaction, closing a scoped session on exit. Thus, here we
should end the current transaction (via rollback) and add the History objects to a new
session, as the previous one will be closed.