Ensure request-scoped SA session is always closed
In some cases we never reach the end of the handle_request() method; as
a result, in those cases the request-scoped SQLAlchemy session is never closed.
With autocommit=False this leads to database transactions being stuck in
idle state, which will very soon raise the psycopg2 "too many clients"
error.

Here we set/unset the request_id on the scoped session within a
try/finally block, thus ensuring the session will be always closed.