Add write intent option to File Sources
This options helps to identify when we are browsing a file source with the "intent to write" to it.

This is helpful to avoid listing those elements (directories/records) that might not be writable (protected, read-only, etc.) even if the file source itself is writable.

The plugin implementation should handle this option accordingly if necessary.