Fix FilesSourcePluginList schema
The common field `browsable` was missing and only when browsable is True we require the `uri_root`.

We can also drop the `Extra.allow` config in FilesSourcePlugin model because the specific fields for different file sources are not returned in the API response since `plugins_to_dict` is called with for_serialization=False by default.