Ignore errors with user-set job resources
Fixes
https://sentry.galaxyproject.org/share/issue/735cef91223e42888ff9bc54ec93c2cb/,
which is a problem that happens when a workflow with resource parameters
is imported from another Galaxy instance, or when we change the
available resource parameters and the selection becomes invalid.

Medium-term we need to stop doing this completely IMO and work with more
abstract job requirements IMO, as this is a big barrier to transfer
workflows from instance to instance and having them just work.