Pin sphinx-rtd-theme to <2 to revert update to pre-release version
Due to a bug in poetry ( https://github.com/python-poetry/poetry/issues/8194 ),
the 2.0.0rc2 pre-release of sphinx-rtd-theme was selected to allow
the update of docutils to 0.19, instead of sticking with the current
version for both.
This breaks the doc build with this error:

```
writing output... [  0%] admin/apache

Theme error:
An error happened in rendering the page admin/apache.
Reason: AttributeError("'str' object has no attribute 'attributes'")
```