Workaround for XML nodes of job resource parameters losing their children
This bug was spotted because the job resource parameters selector broke on usegalaxy.eu. For the tool keras_train_and_eval, the option "Use GPU resources" shows up, but no drop-down menu is available and no items can be selected.

The options are available as children of the XML nodes tagged with "param" within the scope of `galaxy.jobs.parse_resource_parameters`. They are then returned and later retrieved within the scope of `galaxy.util.parse_resource_parameters`, but by then, the children of the XML node (the options) are gone.

See https://github.com/galaxyproject/galaxy/issues/16727 for more details.