Change log level for duplicate data table entries to warning
Fixes https://github.com/galaxyproject/galaxy/issues/16987,
though there's probably something to be done for loc files
being loaded from tools. At the minimum this is a versioning
issue, if a new version of featureCounts comes with a new
built-in reference then an old version shouldn't use it,
and vice-versa. Seems like a hard problem to solve.