Fixes for job metrics Galaxy markdown component.
So https://github.com/galaxyproject/galaxy/pull/10121 added back in the title header to the component which is only designed for the dataset details page for an individual dataset - the original implementation striped out the title for that reason.

The target might an interesting step of a workflow, a problematic tool, a complex resulting dataset - so trying to come up with a title automically is a little challenging. My thought was therefore to keep the component rendered as simply as possible - since the page or report writer will be the one that has the context for how the component should be titled. It needs a prefix and so the big h2 "Job Metrics" is not useful or wanted I think.

https://github.com/galaxyproject/galaxy/pull/10121 did correctly preserve the removal of the title for the job parameters component - so I think this getting added back in was just an oversight.