Disallow empty inputs on required input parameters in simplified workflow form
This is another attempt at
https://github.com/galaxyproject/galaxy/issues/13220, but taking into
account
https://github.com/galaxyproject/galaxy/issues/13220#issuecomment-1068054414
by limiting this to the simplified workflow form, where we can assume
input parameter optionality to have the correct meaning.

This makes it harder to submit a mandatory text parameter that can be
empty, but that is (at this point at least) a very fringe requirement.

If really needed you can make the parameter optional (and use pick_value
to fill in an empty string if null is not OK as a value).