Replace discouraged Mambaforge with Miniforge3
As mentioned on https://github.com/conda-forge/miniforge :

With the release of Miniforge3-23.3.1-0, the packages and configuration of
Mambaforge and Miniforge3 are now identical. The only difference between the
two is the name of the installer and, subsequently, the default installation
directory.

Given its wide usage, there are no plans to deprecate Mambaforge. If at some
point we decide to deprecate Mambaforge, it will be appropriately announced
and communicated with sufficient time in advance.

As of September 2023, the new usage of Mambaforge is thus discouraged. Bug
reports specific to Mambaforge will be closed as won't fix.