add show recent functionality to multiview, uniform multi history view
This adds the option of showing 4 latest histories in Multiview,
instead of always tracking pinned histories. Once the user decides
to track pinned histories, we do so, until the user "resets" it to
recent histories mode.

Also organized the multiple histories in Multiview to show up evenly,
with the `HistoryDetails` taking up a fixed space at the top.