for "panel activities" only toggle sideBar if `route !== activity.to`
Alter the way the `ActivityBar` treats "panel activities";
i.e.: activities that have a sideBar associated,
so that if the sidebar is active but the current route isn't `activity.to`, then keep sidebar active and
just change route.

Also, if `activity.to` is current route, always keep activity's `is-active` prop = true regardless of sidebar for the activity being active or not.