add element_identifier and ext to inputs config file export
element identifiers are crucial for collections and need to be
exposed in the inputs configfile export. it's too hacky to
parse the identifier from the staging path (plus: the staging
path is sanitised).

I suggest to make the element_identifier a list, because it is
for collections (there the first list elements are collection
identifiers and the last the element identifier of the dataset),
such that users (tool developers) do not need to check if it is
a list or a string.