[23.0] Fix `datasetStore.saveDatasets` bug
When we update the datasets in the `datasetStore`, we can sometimes set multiple properties to undefined.
This happens because a dataset in the store fetched from `/api/datasets/${id}` has a keys mismatch from a dataset coming from `/api/histories/${historyId}/contents` in `watchHistory`, where the latter has lesser keys.
To address this, I replaced 'localItem' with 'item' in the loop that updates the keys, ensuring the correct properties are retained during the store update.