allow etree.XMLParser to resolve external entities
We use XML external (file) entities to include .rst files for the help section in tool XMLs. U ntil recently this wasn't a problem, but the default in lxml was recently changed and there is no way in planemo linting or testing to set this flag for the XMLParser. 

This change mainly makes parse_xml() use the old default.