[24.0] Fix history panel arrow navigate by id bug
For navigating history with arrow keys, we were tracking the next item by the `item.id` but items can have the same id and therefore, we were navigating across items. Changed it to `item.hid` instead because that is unique for the items visible for the current `historyId` and `filterText`.