add a first and last item tracker to `SelectedItems`, remove Shift+Ctrl
Modified `SelectedItems` so that for range select, we track the first and last item in a range, along with the anchor. This way, we can persist with a selected range even if the user deselects a few items within it, since we keep track of the beginning and end of the range.

Also, added a CSS indicator to `ContentItem`s which actually indicates which item if the Shift+click range select anchor.
This also removes the `Shift+Ctrl+Click` multiple range select.