[24.0] Do not save workflow on Run without user confirmation
This adds a modal when the user runs a workflow with changes, that asks for user confirmation on whether to proceed without saving changes or to save changes and then proceed. Currently, we _always_ save the workflow when it is run (even if there are no changes at all).

Fixes https://github.com/galaxyproject/galaxy/issues/17903