Update _thanks.rst to include new socials and remove inactive ones
Updating the _thanks.rst that is associated with the User Release Notes to reflect Galaxy's new social media presence.

Galaxy is no longer active on Twitter and has joined Bluesky and LinkedIn. The new 'thanks' message has been updated to reflect this with the appropriate links to the new socials.